Isolate audit file implementation
See #11
